There should be no need for a cast here
...presumably added back in the day to avoid warnings about signed vs. unsigned
comparisons with old compiler versions

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135292
Tested-by: Jenkins

+++ comphelper/qa/container/comphelper_ifcontainer.cxx
@@ -114,7 +114,7 @@ namespace comphelper_ifcontainer
 
             CPPUNIT_ASSERT_EQUAL_MESSAGE("query contents",
                                    nTests, static_cast<int>(aElements.size()));
-            if (static_cast<int>(aElements.size()) == nTests)
+            if (aElements.size() == nTests)
             {
                 for (i = 0; i < nTests; i++)
                 {
